requirements.txt
psycopg2-binary
python-dotenv
requests
selenium
webdriver-manager
